PACKET_INTERCEPTION/requirements.txt
scapy
netfilterqueue
python-dotenv
netifaces
zstandard
# nfqp3
lz4